# apps/backend/articles/views/articles.py
from django.db.models import OuterRef, Subquery
from django_filters import rest_framework as filters
from drf_std_response import EnvelopeMixin
from rest_framework import status
from rest_framework.decorators import action
from rest_framework.permissions import IsAuthenticated
from rest_framework.viewsets import ModelViewSet, ReadOnlyModelViewSet

from core.utils.permissions import is_moderator

from ..filters import ArticleFilter
from ..models import Article, ArticleEvent, ArticlePublication, ArticleSnapshot
from ..permissions import (
    ArticleEventPermission,
    AuthorOnly,
    ModeratorOnly,
)
from ..serializers import (
    ArticleActionResponseSerializer,
    ArticleEventSerializer,
    ArticlePublicationSerializer,
    ArticleReadSerializer,
    ArticleSnapshotSerializer,
    ArticleWriteSerializer,
    ImageUploadResponseSerializer,
    ImageUploadSerializer,
)
from ..services.articles import (
    approve,
    create_article,
    reject,
    save_draft,
    soft_delete,
    submit,
    unpublish,
    withdraw,
)


class ArticleViewSet(EnvelopeMixin, ModelViewSet):
    queryset = Article.objects.select_related("author", "source")
    filter_backends = [filters.DjangoFilterBackend]
    filterset_class = ArticleFilter
    openapi_request_serializer_mapping = {
        "upload_images": ImageUploadSerializer,
        "submit": None,
        "withdraw": None,
        "approve": None,
        "reject": None,
        "unpublish": None,
        "trash": None,
    }
    openapi_response_serializer_mapping = {
        "upload_images": ImageUploadResponseSerializer,
        "submit": ArticleActionResponseSerializer,
        "withdraw": ArticleActionResponseSerializer,
        "approve": ArticleActionResponseSerializer,
        "reject": ArticleActionResponseSerializer,
        "unpublish": ArticleActionResponseSerializer,
        "trash": ArticleActionResponseSerializer,
    }

    permission_class_mapping = {
        'create': [IsAuthenticated],
        'list': [AuthorOnly],
        'retrieve': [AuthorOnly],
        'update': [AuthorOnly],
        'partial_update': [AuthorOnly],
        'destroy': [AuthorOnly],
        'upload_images': [AuthorOnly],
        'submit': [AuthorOnly],
        'withdraw': [AuthorOnly],
        'approve': [ModeratorOnly],
        'reject': [ModeratorOnly],
        'unpublish': [ModeratorOnly],
        'trash': [AuthorOnly],
    }

    def get_serializer_class(self):
        # Author - Write Serializer | Moderators - Read Serializer
        self.action: str
        if self.action in ('create', 'update', 'partial_update'):
            return ArticleWriteSerializer
        return ArticleReadSerializer

    def get_permissions(self):
        self.action: str
        self.permission_classes = self.permission_class_mapping[self.action]

        return [permission() for permission in self.permission_classes]

    def get_queryset(self):
        # Only authors can see his/her articles
        user = self.request.user
        queryset = super().get_queryset()

        if not user or user.is_anonymous:
            return queryset.none()

        queryset = queryset.filter(author=user)
        last_snapshot_id = (
            ArticleSnapshot.objects
            .filter(article_id=OuterRef("pk"))
            .order_by("-created_at")
            .values("id")[:1]
        )
        publication_id = ArticlePublication.objects.filter(article_id=OuterRef("pk")).values("id")

        return queryset.annotate(
            last_snapshot_id=Subquery(last_snapshot_id),
            publication_id=Subquery(publication_id),
        )

    def create(self, request, *args, **kwargs):
        input_serializer = ArticleWriteSerializer(
            data=request.data,
            context=self.get_serializer_context(),
        )
        input_serializer.is_valid(raise_exception=True)

        article = create_article(
            actor=self.request.user,
            markdown=input_serializer.validated_data.get("markdown"),
        )

        output_serializer = ArticleReadSerializer(
            instance=article,
            context=self.get_serializer_context(),
        )

        return self.format_success_response(
            message="created",
            code='created',
            data=output_serializer.data,
            status_code=status.HTTP_201_CREATED,
        )

    def update(self, request, *args, **kwargs):
        partial = kwargs.pop('partial', False)
        article = self.get_object()
        input_serializer = ArticleWriteSerializer(
            article,
            data=request.data,
            partial=partial,
            context=self.get_serializer_context(),
        )
        input_serializer.is_valid(raise_exception=True)

        article = save_draft(
            article_id=article.id,
            actor=request.user,
            title=input_serializer.validated_data.get("title"),
            markdown=input_serializer.validated_data.get("markdown"),
        )

        output_serializer = ArticleReadSerializer(
            instance=article,
            context=self.get_serializer_context(),
        )

        return self.format_success_response(
            message="saved",
            code='saved',
            data=output_serializer.data,
            status_code=status.HTTP_200_OK,
        )

    def partial_update(self, request, *args, **kwargs):
        kwargs['partial'] = True
        return self.update(request, *args, **kwargs)

    @action(detail=False, methods=['post'], url_path='images')
    def upload_images(self, request):
        serializer = ImageUploadSerializer(data=request.data)
        serializer.is_valid(raise_exception=True)

        result = serializer.save()

        return self.format_success_response(
            message="uploaded",
            code='uploaded',
            data=result,
            status_code=status.HTTP_200_OK,
        )

    @action(detail=True, methods=['post'])
    def submit(self, request, pk=None):
        result = submit(article_id=pk, actor=request.user)
        output_serializer = ArticleActionResponseSerializer(instance=result)

        return self.format_success_response(
            message="submitted",
            code='submitted',
            data=output_serializer.data,
            status_code=status.HTTP_200_OK,
        )

    @action(detail=True, methods=['post'])
    def withdraw(self, request, pk=None):
        result = withdraw(article_id=pk, actor=request.user)
        output_serializer = ArticleActionResponseSerializer(instance=result)

        return self.format_success_response(
            message="withdrawn",
            code='withdrawn',
            data=output_serializer.data,
            status_code=status.HTTP_200_OK,
        )

    @action(detail=True, methods=['post'])
    def approve(self, request, pk=None):
        result = approve(article_id=pk, actor=request.user)
        output_serializer = ArticleActionResponseSerializer(instance=result)

        return self.format_success_response(
            message="approved",
            code='approved',
            data=output_serializer.data,
            status_code=status.HTTP_200_OK,
        )

    @action(detail=True, methods=['post'])
    def reject(self, request, pk=None):
        result = reject(article_id=pk, actor=request.user)
        output_serializer = ArticleActionResponseSerializer(instance=result)

        return self.format_success_response(
            message="rejected",
            code='rejected',
            data=output_serializer.data,
            status_code=status.HTTP_200_OK,
        )

    @action(detail=True, methods=['post'])
    def unpublish(self, request, pk=None):
        result = unpublish(article_id=pk, actor=request.user)
        output_serializer = ArticleActionResponseSerializer(instance=result)

        return self.format_success_response(
            message="unpublished",
            code='unpublished',
            data=output_serializer.data,
            status_code=status.HTTP_200_OK,
        )

    @action(detail=True, methods=['post'], url_path="delete")
    def trash(self, request, pk=None):
        """Soft delete the article"""
        result = soft_delete(article_id=pk, actor=request.user)
        output_serializer = ArticleActionResponseSerializer(instance=result)

        return self.format_success_response(
            message="deleted",
            code='deleted',
            data=output_serializer.data,
            status_code=status.HTTP_200_OK,
        )


class ArticlePublicationViewSet(EnvelopeMixin, ReadOnlyModelViewSet):
    queryset = ArticlePublication.objects.select_related("article").prefetch_related("versions")
    serializer_class = ArticlePublicationSerializer
    permission_classes = [IsAuthenticated]

    def get_queryset(self):
        from comments.querysets import with_article_publication_comment_count
        from reactions.querysets import with_article_publication_reaction_summary

        queryset = with_article_publication_reaction_summary(
            super()
            .get_queryset()
            .filter(article__status=Article.ArticleStatus.PUBLISHED)
            .order_by("-published_at", "-created_at"),
            user=self.request.user,
        )
        return with_article_publication_comment_count(queryset)


class ArticleSnapshotViewSet(EnvelopeMixin, ReadOnlyModelViewSet):
    queryset = ArticleSnapshot.objects.all()
    serializer_class = ArticleSnapshotSerializer
    permission_classes = [ModeratorOnly]
    openapi_response_serializer_mapping = {
        "pending_ones": ArticleSnapshotSerializer(many=True),
    }

    @action(detail=False, methods=['get'])
    def pending_ones(self, request):
        """
        Return not moderated article snapshots
        """
        queryset = ArticleSnapshot.objects.filter(moderation_status=ArticleSnapshot.SnapshotStatus.PENDING)
        serializer = self.get_serializer(queryset, many=True)

        return self.format_success_response(
            message="pending articles listed",
            code='listed',
            data=serializer.data,
            status_code=status.HTTP_200_OK,
        )


class ArticleEventReadViewset(EnvelopeMixin, ReadOnlyModelViewSet):
    queryset = ArticleEvent.objects.all()
    permission_classes = (ArticleEventPermission,)
    serializer_class = ArticleEventSerializer

    def get_queryset(self):
        user = self.request.user
        if is_moderator(user):
            return ArticleEvent.objects.all()
        return ArticleEvent.objects.filter(article__author=user)