# apps/backend/openapi/v1.yaml
openapi: 3.0.3
info:
  title: AlienCommons API
  version: 1.0.0
  description: HTTP API for the AlienCommons community platform.
paths:
...
  /v1/articles/{id}/delete/:
    post:
      operationId: articles_delete_create
      description: Soft delete the article
      parameters:
        - in: path
          name: id
          schema:
            type: string
            format: uuid
            description: The UUID of the object
          required: true
      tags:
        - articles
      security:
        - cookieAuth: []
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - message
                  - code
                  - data
                  - errors
                  - meta
                properties:
                  success:
                    type: boolean
                    enum:
                      - true
                  message:
                    type: string
                  code:
                    type: string
                  data:
                    $ref: "#/components/schemas/ArticleActionResponse"
                  errors:
                    type: array
                    nullable: true
                    items:
                      type: object
                      required:
                        - code
                        - message
                        - field
                      properties:
                        code:
                          type: string
                        message:
                          type: string
                        field:
                          type: string
                          nullable: true
                  meta:
                    type: object
                    required:
                      - request_id
                      - timestamp
                    properties:
                      request_id:
                        type: string
                        nullable: true
                      timestamp:
                        type: string
                        format: date-time
                        nullable: true
          description: ""
...
  /v1/articles/images/:
    post:
      operationId: articles_images_create
      description: Wrap successful DRF responses in the standard response envelope.
      tags:
        - articles
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImageUploadRequest"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/ImageUploadRequest"
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/ImageUploadRequest"
        required: true
      security:
        - cookieAuth: []
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - message
                  - code
                  - data
                  - errors
                  - meta
                properties:
                  success:
                    type: boolean
                    enum:
                      - true
                  message:
                    type: string
                  code:
                    type: string
                  data:
                    $ref: "#/components/schemas/ImageUploadResponse"
                  errors:
                    type: array
                    nullable: true
                    items:
                      type: object
                      required:
                        - code
                        - message
                        - field
                      properties:
                        code:
                          type: string
                        message:
                          type: string
                        field:
                          type: string
                          nullable: true
                  meta:
                    type: object
                    required:
                      - request_id
                      - timestamp
                    properties:
                      request_id:
                        type: string
                        nullable: true
                      timestamp:
                        type: string
                        format: date-time
                        nullable: true
          description: ""
...
components:
  schemas:
...
    ArticleActionResponse:
      type: object
      properties:
        event_type:
          type: integer
        actor_id:
          type: string
          format: uuid
        article_id:
          type: string
          format: uuid
        article_snapshot_id:
          type: string
          format: uuid
          nullable: true
        event_id:
          type: string
          format: uuid
        event_type_display:
          type: string
          readOnly: true
      required:
        - actor_id
        - article_id
        - article_snapshot_id
        - event_id
        - event_type
        - event_type_display
...
    ImageUploadRequest:
      type: object
      properties:
        image:
          type: string
          format: binary
      required:
        - image
    ImageUploadResponse:
      type: object
      properties:
        url:
          type: string
          readOnly: true
        path:
          type: string
          readOnly: true
        name:
          type: string
          readOnly: true
      required:
        - name
        - path
        - url
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid